requirements.txt
torch
torchvision
numpy
opencv-python
Pillow
albumentations
tqdm
